move table component to components
